--- src/components/tools/Stroke.ts
@@ -1,4 +1,0 @@
-export type Stroke = {
-  attributes: Record<string, string | number>;
-  type: string;
-};
