+++ src/components/whiteboard/Pen.ts
