+++ src/components/whiteboard/Point.ts
@@ -1,0 +1,4 @@
+export type Point = {
+  x: number;
+  y: number;
+};
